requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
python-decouple==3.8
